+++ app/src/features/rules/screens/rulesList/components/RulesList/components/GettingStarted/gettingStarted.scss
