--- app/src/features/rules/screens/rulesList/components/RulesList/components/GettingStarted/v1/gettingStarted.scss
@@ -1,59 +1,0 @@
-.v1.getting-started-container {
-  padding: 2.5rem;
-  min-height: 82vh;
-
-  .getting-started-header-container {
-    margin-bottom: 1rem;
-  }
-
-  .getting-started-header {
-    font-size: 16px;
-  }
-
-  .getting-started-video-container {
-    padding: 56.25% 0 0 0;
-    position: relative;
-    -webkit-box-shadow: 0px 7px 16px 1px rgba(15, 15, 15, 1);
-    -moz-box-shadow: 0px 7px 16px 1px rgba(15, 15, 15, 1);
-    box-shadow: 0px 7px 16px 1px rgba(15, 15, 15, 1);
-    border-radius: 20px;
-  }
-
-  .getting-started-video-container video {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 100%;
-    border: 1px solid var(--border-dark);
-    border-radius: var(--border-radius-md);
-  }
-  .getting-started-video-container iframe {
-    border-radius: inherit;
-  }
-
-  .getting-started-actions {
-    margin-top: 1.25rem;
-    margin-bottom: 1.75rem;
-  }
-  .getting-started-btns-wrapper {
-    display: flex;
-    gap: 12px;
-  }
-  .getting-started-subtitle {
-    margin-bottom: 8px;
-  }
-
-  .getting-started-signin-link {
-    color: #ffffff;
-    border: none;
-    cursor: pointer;
-    padding: 1px 6px;
-    background-color: var(--component-background);
-    border-radius: var(--border-radius-sm);
-  }
-
-  .getting-started-signin-link:hover {
-    background-color: #3c3d42;
-  }
-}
